+++ Settings.h
@@ -1,10 +1,10 @@
 // #############################################################################
 //
 // # Name       : Settings
-// # Version    : 1.6
+// # Version    : 1.7
 //
 // # Author     : Juan L. Perez Diez <ender.vs.melkor at gmail>
-// # Date       : 13.01.2016
+// # Date       : 20.06.2016
 //
 // # Description: Settings class for Huertomato
 // # Stores all the system's current settings. Its in charge of reading and storing in EEPROM 
@@ -91,8 +91,8 @@ class Settings {
     boolean setFloodMinute(const uint8_t);
     boolean setPHalarmUp(const float);
     boolean setPHalarmDown(const float);
-    boolean setECalarmUp(const uint16_t);
-    boolean setECalarmDown(const uint16_t);
+    boolean setECalarmUp(const float);
+    boolean setECalarmDown(const float);
     boolean setWaterAlarm(const uint8_t);
     boolean setNightWatering(const boolean);
 	boolean setLightThreshold(const uint16_t);
@@ -129,8 +129,8 @@ class Settings {
     uint8_t getFloodMinute() const;
     float getPHalarmUp() const;
     float getPHalarmDown() const;
-    uint16_t getECalarmUp() const;
-    uint16_t getECalarmDown() const;
+    float getECalarmUp() const;
+    float getECalarmDown() const;
     uint8_t getWaterAlarm() const;
     boolean getNightWatering() const;
 	uint16_t getLightThreshold() const;
@@ -188,8 +188,8 @@ class Settings {
     //Sensor Alarms
     float _phAlarmUp;
     float _phAlarmDown;
-    uint16_t _ecAlarmUp;
-    uint16_t _ecAlarmDown;
+    float _ecAlarmUp;
+    float _ecAlarmDown;
     uint8_t _waterAlarm;
     //Water at night - Informs the system that watering timers are stopped for the night
     boolean _nightWatering;
